Create a custom hook, useRefs
